+++ react/__tests__/MiniCart.test.js
@@ -125,4 +125,32 @@ describe('<MiniCart /> component', () => {
       expect.objectContaining({ title: expect.any(String) })
     )
   })
+
+  it('should not show item quantity if there are no items in cart', () => {
+    const {queryByTestId} = render(
+      <MiniCart type="sidebar" hideContent={false} showTotalItemsQty={false}/>
+    )
+    expect(queryByTestId('item-qty')).toBeNull()
+  })
+
+  it('should show the quantity of different items in cart', async () => {
+    const {getByTestId} = render(
+      <MockedProvider mocks={mocks}>
+        <MiniCart type="sidebar" hideContent={false} showTotalItemsQty={false}/>
+      </MockedProvider>
+    )
+    await resolveApolloQueries()
+    expect(getByTestId('item-qty').textContent).toBe('1')
+  })
+
+  it('should show the quantity of different items in cart', async () => {
+    const {getByTestId} = render(
+      <MockedProvider mocks={mocks}>
+        <MiniCart type="sidebar" hideContent={false} showTotalItemsQty={true}/>
+      </MockedProvider>
+    )
+    await resolveApolloQueries()
+    expect(getByTestId('item-qty').textContent).toBe('2')
+  })
+
 })